+++ MedixCollege/Controllers/HomeController.cs
@@ -335,9 +335,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -491,9 +491,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -661,9 +661,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -831,9 +831,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -1001,9 +1001,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -1171,9 +1171,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -1341,9 +1341,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["Telephone"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
@@ -1387,9 +1387,9 @@ namespace MedixCollege.Controllers
                 return RedirectToRoute("ThankYou");
             }
 
-            int phoneNumber = 0;
+            double phoneNumber = 0;
 
-            int.TryParse(Helpers.Helpers.GetNumbers(fc["PhoneNumber"]), out phoneNumber);
+            double.TryParse(Helpers.Helpers.GetNumbers(fc["PhoneNumber"]), out phoneNumber);
 
             if (phoneNumber == 0)
             {
